% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeOverlap.R
\name{arbtr_ol}
\alias{arbtr_ol}
\title{Title performing arbitrary overlap calculation among a set of genomic ranges}
\usage{
arbtr_ol(xset, min.gapwidth = 0L, mode = c("interval", "base"), ...)
}
\arguments{
\item{xset}{a list of GenomicRanges object}

\item{min.gapwidth}{minimum gap to merge hits, default is 0}

\item{mode}{decide counting overlapping intervals or overlapping bases as
intersection size, default is interval}

\item{...}{}
}
\value{
An overlap object containing a overlap matrix and a list containing
corresponding overlapping intervals
}
\description{
provide various options to define overlapping
}
\examples{
 \dontrun{
  gr1 <- IRanges(c(15,6,44),c(20,8,50))
  gr2 <- IRanges(c(8,18,20),c(16,22,24))

  #running under default parameters
  grlist <- list(gr1,gr2)
  res <- arbtr_ol(grlist)
  res[[0]]
  res[[1]]

  #alternating rules defining intersection
  res <- arbtr_ol(grlist,minoverlap = 3L)
  res[[0]]
  res[[1]]

  #merging overlapping regions with distance less than n bases counts
  res <- arbtr_ol(grlist,min.gapwidth = 3L)
  res[[0]]
  res[[1]]

  #counting overlapping bases instead of number of overlapping intervals,
  #which affects the calculation of intersection sizes
  res <- arbtr_ol(grlist,mode = 'base')
  res[[0]]
  res[[1]]
 }

}
