% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bed2gr}
\alias{bed2gr}
\title{Title transform bed-format file into genomic ranges object}
\usage{
bed2gr(file)
}
\arguments{
\item{file}{can be a string representing file path or a data frame object}
}
\value{
a genomic ranges object
}
\description{
Title transform bed-format file into genomic ranges object
}
