% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gr2df}
\alias{gr2df}
\title{Title transform genomic ranges object into data frame which can be directly
saved into bed format file}
\usage{
gr2df(gr)
}
\arguments{
\item{gr}{single genomic range object or genomic range list object or just
a list}
}
\value{
a data frame
}
\description{
Title transform genomic ranges object into data frame which can be directly
saved into bed format file
}
