% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectplot.R
\name{olm2upset}
\alias{olm2upset}
\title{Title visualize multiple intersection by upset plot}
\usage{
olm2upset(oldf, outputfig = NULL, method = "UpSetR", ...)
}
\arguments{
\item{oldf}{overlap matrix demonstrating overlapping counts of
each combinations, can be provided by other methods}

\item{outputfig}{file path for saving figure, if one don't want to
save it, keep it as default}

\item{method}{options of visualization. Currently only UpSetR is supported}

\item{...}{}
}
\value{
plotting object
}
\description{
Title visualize multiple intersection by upset plot
}
