% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectplot.R
\name{olm2venn}
\alias{olm2venn}
\title{Title visualizing multiple intersection by venn plot}
\usage{
olm2venn(oldf, outputfig = NULL, method = c("gplots", "ggvenn"), ...)
}
\arguments{
\item{oldf}{overlap matrix demonstrating overlapping counts of
each combination, can be provided by other methods}

\item{outputfig}{file path for saving figure, if one don't want to
save it, keep it as default}

\item{method}{options of visualization. Currently gplots and ggvenn are
supported}

\item{...}{}
}
\value{
plotting object
}
\description{
Title visualizing multiple intersection by venn plot
}
